Main contig grabber code
Multifunctional, able to operate on command-line and as a CMD-window operated program. Also has a helpful effort function which allows for partial sequence ID matches to be performed, which is helpful when dealing with programs like BLAST/MMseqs2 which split sequence IDs at space characters. Thus, if you have a list of sequence IDs pulled out of a BLAST table, so long as they uniquely partially map back to a FASTA sequence ID, we can reassociate these results.