commonStart update 1
This version doesn't work well. It's worse than before because it tries to force alignments to match things in the commonStarts list. The sorting mechanism needs reprioritising OR some sort of offset measure to prevent excessive shortening OR some sort of "homology score" to only trim sequences if it doesn't align well to other sequences.