AlignIO fixed codon translation bugs and better control
The alignments should come out properly now (were some odd instances with bugs) and also we've dropped the findBestFrame parameter in favour of inferring this based on strand and frame args. It makes use of the flexible translation for better functioning when we do / do not know what translation strand/frame we want.