3_polish fix bug in add_codon_numbers and reuse frames
We can reuse the frames from the polishing function to possibly get more accurate frame prediction especially when it's unclear which frame is best and we end up using BLAST against the transcriptome to figure it out.

Also, WTF? All the codon numbers would be wrong prior to this fix. max(Counter) does not work like I thought it did.