filter_vcf.py use VCFIO and add new filters
Easier to work with VCFs with the new code. Added new filters to remove samples not in the pops file, replace a sample's genotype with ambiguity where its DP is too low and stuff. There might be a bug or two in the code still since it needs proper testing but it might be okay already?