GFF3IO .update_id and .reformat_id
Make a clearer distinction between the goals of these functions. The first is to just change a feature.ID and have its immediate children point to that with their .Parent field. reformat_id will actually generate new IDs for all children downstream of the feature being renamed.

This change is (hopefully) applied to the other codes that made use of this function. It's clear how sloppily it's implemented now, but it's kind of too monolithic to change or refactor without significant time investment...