predict_chimeric_amplicons imperfect revision to align
This method sort of works, in that it prevents some erroneous chimera classification. But, it causes issues by trimming parts from a sequence that should align against the full MSA but don't align against the best reference hit. I probably need to perform MUSCLE twice to work around this which sucks since that's the main time consuming step