gene_haplotyping_pipe readgroup resume
Treat each BAM file separately as we do with the subsetting and indexing, to make it easier to add more samples into an analysis. It still means the user will need to delete any existing flags and files for the downstream steps, if they want to update an analysis, but it is what it is.