ZS_ORF handle concealed positions
Gaps or Ns at the start and end of a sequence can "conceal" stop codons, and its important for the lineChart scoring that we factor those in somehow. Penalising concealed positions on its own ruins the behaviour of the algorithm, but giving scores where a stop codon probably is also degrades its performance.